requirements.txt
gradio>=4.0.0
python-dotenv>=1.0.0
google-generativeai>=0.3.0
openai-whisper>=20231117
TTS>=0.22.0
numpy>=1.23.0
opencv-python>=4.7.0.72
ffmpeg-python>=0.2.0
